added a map implementation for the server scheduler repositories

+++ spring-batch-lightmin-server/spring-batch-lightmin-server-scheduler/spring-batch-lightmin-server-scheduler-core/src/main/java/org/tuxdevelop/spring/batch/lightmin/server/scheduler/service/SchedulerConfigurationService.java
@@ -4,6 +4,7 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.tuxdevelop.spring.batch.lightmin.server.scheduler.repository.SchedulerConfigurationRepository;
 import org.tuxdevelop.spring.batch.lightmin.server.scheduler.repository.domain.SchedulerConfiguration;
+import org.tuxdevelop.spring.batch.lightmin.server.scheduler.repository.domain.SchedulerValidationException;
 import org.tuxdevelop.spring.batch.lightmin.server.scheduler.repository.exception.SchedulerConfigurationNotFoundException;
 
 public class SchedulerConfigurationService {
@@ -16,15 +17,20 @@ public class SchedulerConfigurationService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public SchedulerConfiguration save(final SchedulerConfiguration schedulerConfiguration) {
-
-        //TODO: validate
-        return this.schedulerConfigurationRepository.save(schedulerConfiguration);
+        if (schedulerConfiguration == null) {
+            throw new SchedulerValidationException("schedulerConfiguration must not be null");
+        } else {
+            return this.schedulerConfigurationRepository.save(schedulerConfiguration);
+        }
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
     public void delete(final SchedulerConfiguration schedulerConfiguration) {
-        //TODO: validate
-        this.schedulerConfigurationRepository.delete(schedulerConfiguration.getId());
+        if (schedulerConfiguration == null) {
+            throw new SchedulerValidationException("schedulerConfiguration must not be null");
+        } else {
+            this.schedulerConfigurationRepository.delete(schedulerConfiguration.getId());
+        }
     }
 
     @Transactional(readOnly = true)
